# apps/unstructured/docs/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PrimeData Backend API",
    "description": "Enterprise data quality and AI readiness platform with document processing, quality validation, and vector search capabilities",
    "version": "1.0.0",
    "contact": {
      "name": "PrimeData Support",
      "url": "https://primedata.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.primedata.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "ChunksResponse": {
        "type": "object",
        "properties": {
          "total_count": {"type": "integer"},
          "offset": {"type": "integer"},
          "limit": {"type": "integer"},
          "chunks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {"type": "string"},
                "text": {"type": "string"},
                "source": {"type": "string", "description": "Source file name"},
                "page_number": {"type": "integer"},
                "chunk_index": {"type": "integer"},
                "section": {"type": "string"},
                "quality_scores": {
                  "type": "object",
                  "properties": {
                    "coherence": {"type": "number", "format": "float"},
                    "noise": {"type": "number", "format": "float"},
                    "sentence_complete": {"type": "boolean"}
                  }
                },
                "relevance_score": {"type": "number", "format": "float", "description": "Vector search relevance score"},
                "created_at": {"type": "string", "format": "date-time"}
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/products/{product_id}/chunks": {
      "get": {
        "tags": ["Chunks"],
        "summary": "Search chunks (vector search)",
        "operationId": "getChunks",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "schema": {"type": "string"}
          },
          {
            "name": "version",
            "in": "query",
            "schema": {"type": "integer"}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 50}
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {"type": "integer", "default": 0}
          }
        ],
        "responses": {
          "200": {
            "description": "Chunks search results",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChunksResponse"}}}
          }
        }
      }
    },
...
    "/api/v1/acl": {
      "post": {
        "tags": ["Governance"],
        "summary": "Create ACL permission",
        "operationId": "createACL",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {"type": "string"},
                  "resource_type": {"type": "string"},
                  "resource_id": {"type": "string", "format": "uuid"},
                  "permission": {"type": "string", "enum": ["READ", "WRITE", "ADMIN"]}
                },
                "required": ["user_id", "resource_type", "resource_id", "permission"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "ACL created"}
        }
      },
      "get": {
        "tags": ["Governance"],
        "summary": "List ACLs",
        "operationId": "listACLs",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "List of ACLs"}
        }
      }
    },
...
  },
  "security": [
    {"bearerAuth": []}
  ]
}